requirements.txt
Flask==3.0.3
websocket-client==1.8.0
requests==2.32.3
numpy==2.3.4
scipy==1.16.1